Set enum_kind based on Fortran standard
The enum_kind is now set in the fortran_c_interop module but made
publicly availabe via the client and dataset modules. This differs
from the previous implementation which attempted to set enum_kind
by creating a dummy enum and detecting the kind of it at compile
time. This however, broke compilation with GCC 6.3.0, because
that compiler chain allows for different 'types' of enums (a
departure from the Fortran standard).

This also removes the explicit declaration of the kind in most of
the tests since Fortran assignment should automatically convert
the value to the correct kind that the variable (e.g. err_code)
is declared as.