Implement dataset aggregation for Fortran
Adds in all methods and C interfaces to enable the dataset
aggregation functionality in Fortran. Most of the transformations
and implementations are similar to those done before. The one
departure from the C implementation is that for returning a list
of datasets, we rely on Fortran to allocate the necessary memory
for the list of dataset objects. This vector is then populated
within the C client without allocating a new vector.

This is just an initial implementation. Tests to follow