Change Docker container image to 20.04 LTS
Github Actions fail due to the docker build step relying on
Ubuntu 21.04. This version of Ubuntu has reached end of life and
so the package install step of the workflow was failing. The rest
of the testing for SmartRedis points to 20.04 and so both the
tests and Dockerfile is synced.