Change Docker container image to 20.04 LTS (#276)
Github Actions fail due to the docker build step relying on
Ubuntu 21.04. This version of Ubuntu has reached end of life and
so the package install step of the workflow was failing. The change
in this PR unifies the image used for both the testing and docker
container for SmartRedis.