Framework for dataset conversion (xarray) (#262)
DatasetConverter class with add_metadata_for_xarray 
and transform_to_xarray method. Dataset conversion 
refers to a multi-step workflow where a user can create 
a SmartRedis dataset which can then be retrieved on 
the Python side, and then transformed into a {Dataformat}
object. Dataset conversion support has only been added
for Xarray. Additional metadata is added to an existing 
SmartRedis dataset to add the possibility of conversion 
in the add_metadata_for_xarray method. The 
transform_to_xarray method ingests a Dataset and 
returns a dictionary of  Xarray Dataarrays. 

This PR includes documentation and tests for all 
of the above.