Logging for SmartRedis (#281)
Implement a new logging facility for C, C++, Python, and Fortran clients. Log messages are written to a logging file pointed to by the SR_LOG_FILE environment variable, or stdout if the variable is not set. Log messages are written conditionally if the logging level in the message is less than or equal to the target log level selected in the SR_LOG_LEVEL environment variable, or "INFO" if the variable is not set.

Log messages are prefixed with a source identifier and a timestamp. The source identifier may be a string (Python, C++); or a Client, Dataset, or LogContext object (C++, C, Fortran). (The LogContext is a new object that carries the prefix for logging messages.) A future PR will extend the functionality of both logging modalities to all client languages.

Implement a unified internal API for managing environment variables.

Add logging messages for exceptions, API entry/exit, ThreadPool activity, and for explicit user-invoked logging.

[ committed by @billschereriii ]
[ reviewed by @ashao @mellis13 @MattToast @ankona @al-rigazzi ]