Prevent codecov status checks from failing (#317)
We have wide variability of our total coverage reports even for
non-substantive commits. We suspect that this is because we are
compiling unit tests and the SmartRedis library in cmake's
Release mode which enables -O3 and may, at worse, skip the
execution of some tests and at least confuses the tracer that
marks which lines have been run.

As a workaround for now, we are lowering the target thresholds to
0% so that commits are not marked with the dreaded 'X' due to the
variance that results in code coverage.