Multidb support for SmartRedis (#353)
Add support for Multiple Databases in SmartRedis. This involves a Client class constructor update -- the type of database is now stored in an environment variable rather than being passed in to the Client constructor. The legacy constructor will remain supported for the short term but will be phased out in favor of the new constructor.
Updated all test cases and examples with the new Client class constructor.
Updated documentation.
[ committed by @billschereriii ]
[ reviewed by @ashao @MattToast ]