Update Redis version and fix parallel example
The Redis version now matches the the one we build with SmartSim.
This required a unit test to be updated since the original
dbfilename parameter is now protected.

Additionally, when using the Nvidia compiler chain, the parallel
examples were not working correctly because they were not being
run with mpirun. This may only be a problem with OpenMPI (which is
what NVHPC ships with), whereas Intel and MPICH (which is installed
with GNU during CI/CD) allow MPI applications to run without being
launched by mpirun