Remove unused memory layout constructor parameter for TensorBase. (#479)
The ``TensorBase`` object has a ``SRMemoryLayout`` 
parameter that is not used. The ``SRMemoryLayout`` 
is only used by the child ``Tensor<T>`` class. 
Because all tensors are stored internally has contiguous memory,
there is no need to have it stored as a member variable. 
It has been removed to avoid compiler warnings.

[ committed by @mellis13 ]
[ reviewed by @ashao ]