Fix minor warnings in C++ client (#476)
Warnings were being thrown when compiling the library with some of the
pickiest compiler flags. The offending portions of the code base was 
reworked to ensure that these warnings no longer exist. Additionally, 
the `SR_PEDANTIC` flag was  re-enabled to ensure that defects do
not creep into the codebase. Lastly, all tested compiler chains GCC,
Intel, and NVHPC will compile with the most pedantic flags.

[ committed by @ashao ]
[ reviewed by @mellis13 ]