Fix memory leak in Fortran (#514)
The dataset object, if used in a loop, would leave memory dangling. To
alleviate this, a `final` procedure has been implemented. Fortran
compilers, however, are notoriously bad at detecting when an object goes
out of scope and to destroy them automatically. We thus also provide an
explicit ``destructor`` type procedure.

[ committed by @ashao ]
[ reviewed by @MattToast ]