Fix bug in Fortran dataset destructor (#525)
Fixes a bug in the Fortran client where (on compilers that support
automatic destruction) the program would segfault if the user had
manually destroyed the dataset. This behaviour was tracked down to an
error in the C-client where only the one-level de-referenced pointer was
being set to `NULL` but not the actual double-pointer.

< committed by @ashao >
< reviewed by @al-rigazzi >